Moved settings fragment to it's own file
