+++ app/src/main/res/values/dimens.xml
@@ -52,4 +52,16 @@
     <dimen name="about_page_padding">10dp</dimen>
     <dimen name="header_width">30dp</dimen>
     <dimen name="item_two_line_height">64dp</dimen>
+    <!--
+    Because the window insets on round devices are larger than 15dp, this padding only applies
+    to square screens.
+    -->
+    <dimen name="box_inset_layout_padding">0dp</dimen>
+
+    <!--
+    This padding applies to both square and round screens. The total padding between the buttons
+    and the window insets is box_inset_layout_padding (above variable) on square screens and
+    inner_frame_layout_padding (below variable) on round screens.
+    -->
+    <dimen name="inner_frame_layout_padding">5dp</dimen>
 </resources>
