+++ app/src/main/res/values/strings.xml
@@ -209,5 +209,11 @@
     <string name="enable_permission">Enable Permission</string>
     <string name="button_enable_contact_permission">Enable contacts</string>
     <string name="button_enable_recents_permission">Enable Call Log</string>
+    <string name="title_activity_contact">ContactActivity</string>
+    <!--
+    This string is used for square devices and overridden by hello_world in
+    values-round/strings.xml for round devices.
+    -->
+    <string name="hello_world">Hello Square World!</string>
 
 </resources>
